Don't initialize a remote for an unrelated branch
When importing a merge changeset, if the parent changeset has not been
imported into Git yet and contains changes in multiple branches, and at
least one of those branches doesn't belong to the same branch hierarchy
as the current remote (i.e. they don't have a common root), then we
could pick the unrelated branch instead of the related branch (depending
on the order in which the individual changes are processed).  To avoid
this, we need to ignore changes that don't belong to branches that are
related to the current remote's branch.